port modal component, align the last button example
The modal component is a good example of why we prefer LLM codegen
instead of AST codegen. The implementation of the Vue modal
component is quite complex, which may cause a lot of effort to
implement a robust AST transformer.

The react component has some re-write logic, but at least it pass
the button click example e2e tests.

Now all the button examples is alive and can pass the e2e tests!

Bugs being fixed during this port:

- react-common setup now simulate a emitter factory
- react-common eventBus check callback truthy before push